Added support for SVG, Tiling
Now you can save file in SVG format as well as PNG. More formats can be added, but these are all for now. Tiling is now an option. New UI elements are added to handle tiling. Previous method of tiling was slow. With the method applied here, the frame rate is about 57-58 which is quite good.